Attempt to receive data in order to determin socket closure
Execute the `notify_command` when the socket has been closed instead of
after `dtach` is called. In order to do this attemp to receive IO on the
UNIX socket created by running `dtach -n`. No IO is actually expected
but by listening for IO it is possible to rescue an error when the
socket closes.

Some information about socket interaction was found on the Ruby-Talk
mailing list at
http://blade.nagaokaut.ac.jp/cgi-bin/scat.rb/ruby/ruby-talk/242369